Deliver the sim robot model via `SIM_ROBOT_TRANSFORM`, not the recording
Mirror the real arm: the sim no longer bakes its model into recorded static, which put binary mesh
bytes there and broke the RemoteDataset JSON API. `MujocoSim` emits only `scene_xml`;
`bundled_panda_model` moves to `drivers/roboarm` beside `bundled_franka_model`, and
`SIM_ROBOT_TRANSFORM` supplies the panda URDF + collision meshes + gripper at read time. Recorded and
served episodes are JSON-safe; sim episodes render via `positronic.cfg.analysis sim`.
