+++ positronic/drivers/roboarm/__init__.py
@@ -3,6 +3,7 @@
 This package provides drivers for various robot arms including Franka, Kinova, and SO-101.
 """
 
+import xml.etree.ElementTree as ET
 from abc import ABC, abstractmethod
 from enum import Enum
 from functools import lru_cache
@@ -32,6 +33,29 @@ def bundled_franka_model() -> dict:
         'meshes': {f.name: f.read_bytes() for f in sorted(meshes) if f.suffix == '.stl'},
         'joint_names': [f'joint{i}' for i in range(1, 8)],
         'control_frame': 'end_effector',
+    }
+
+
+@lru_cache(maxsize=1)
+def bundled_panda_model() -> dict:
+    """The bundled simulated Franka panda (arm + hand) for the 3D viewer and offline IK: the panda
+    URDF, its collision meshes, the joint names, the ``end_effector`` control frame — the grasp site
+    where the sim measures ``robot_state.ee_pose`` — and the ``gripper`` spec that slides the fingers
+    from the recorded ``grip`` signal. Supplied to sim datasets by ``SIM_ROBOT_TRANSFORM``, mirroring
+    how ``bundled_franka_model`` backfills the real arm; its ``end_effector`` sits at the simulated
+    grasp site rather than the FR3 physical flange.
+    """
+    urdf_path = Path(__file__).resolve().parents[2] / 'assets' / 'mujoco' / 'panda.urdf'
+    urdf = urdf_path.read_text()
+    mesh_dir = urdf_path.parent / 'assets'
+    mesh_files = {mesh.get('filename') for mesh in ET.fromstring(urdf).iter('mesh')}
+    return {
+        'urdf': urdf,
+        'meshes': {name: (mesh_dir / name).read_bytes() for name in sorted(mesh_files)},
+        'joint_names': [f'joint{i}' for i in range(1, 8)],
+        'control_frame': 'end_effector',
+        # ``grip`` is recorded in [0, 1] (closed→open); each finger slides 0..0.04 m along its axis.
+        'gripper': {'signal': 'grip', 'joints': ['finger_joint1', 'finger_joint2'], 'travel': 0.04},
     }
 
 
@@ -87,4 +111,4 @@ class State(ABC):
         return None
 
 
-__all__ = ['RobotStatus', 'State', 'bundled_franka_model', 'command']
+__all__ = ['RobotStatus', 'State', 'bundled_franka_model', 'bundled_panda_model', 'command']
